tieba_crawler_api/requirements.txt
# Core crawler
aiotieba>=3.3.0
aiohttp>=3.9.0
python-dotenv>=1.0.0

# API server
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
pydantic>=2.6.0
